Split scenario mod and tests.
