+++ rust_dev_preview/examples/rds/src/aurora_scenario/mod.rs
@@ -1,0 +1,632 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+use phf::{phf_set, Set};
+use secrecy::SecretString;
+use std::{collections::HashMap, fmt::Display, time::Duration};
+
+use aws_sdk_rds::{
+    error::ProvideErrorMetadata,
+    operation::create_db_cluster_parameter_group::CreateDbClusterParameterGroupOutput,
+    types::{DbCluster, DbClusterParameterGroup, DbClusterSnapshot, DbInstance, Parameter},
+};
+use sdk_examples_test_utils::waiter::Waiter;
+use tracing::{info, trace, warn};
+
+const DB_ENGINE: &str = "aurora-mysql";
+const DB_CLUSTER_PARAMETER_GROUP_NAME: &str = "RustSDKCodeExamplesDBParameterGroup";
+const DB_CLUSTER_PARAMETER_GROUP_DESCRIPTION: &str =
+    "Parameter Group created by Rust SDK Code Example";
+const DB_CLUSTER_IDENTIFIER: &str = "RustSDKCodeExamplesDBCluster";
+const DB_INSTANCE_IDENTIFIER: &str = "RustSDKCodeExamplesDBInstance";
+
+static FILTER_PARAMETER_NAMES: Set<&'static str> = phf_set! {
+    "auto_increment_offset",
+    "auto_increment_increment",
+};
+
+#[derive(Debug, PartialEq, Eq)]
+struct MetadataError {
+    message: Option<String>,
+    code: Option<String>,
+}
+
+impl MetadataError {
+    fn from(err: &dyn ProvideErrorMetadata) -> Self {
+        MetadataError {
+            message: err.message().map(|s| s.to_string()),
+            code: err.code().map(|s| s.to_string()),
+        }
+    }
+}
+
+impl Display for MetadataError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let display = match (&self.message, &self.code) {
+            (None, None) => "Unknown".to_string(),
+            (None, Some(code)) => format!("({code})"),
+            (Some(message), None) => message.to_string(),
+            (Some(message), Some(code)) => format!("{message} ({code})"),
+        };
+        write!(f, "{display}")
+    }
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub struct ScenarioError {
+    message: String,
+    context: Option<MetadataError>,
+}
+
+impl ScenarioError {
+    pub fn with(message: impl Into<String>) -> Self {
+        ScenarioError {
+            message: message.into(),
+            context: None,
+        }
+    }
+
+    pub fn new(message: impl Into<String>, err: &dyn ProvideErrorMetadata) -> Self {
+        ScenarioError {
+            message: message.into(),
+            context: Some(MetadataError::from(err)),
+        }
+    }
+}
+
+impl std::error::Error for ScenarioError {}
+impl Display for ScenarioError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match &self.context {
+            Some(c) => write!(f, "{}: {}", self.message, c),
+            None => write!(f, "{}", self.message),
+        }
+    }
+}
+
+// Parse the ParameterName, Description, and AllowedValues values and display them.
+#[derive(Debug)]
+pub struct AuroraScenarioParameter {
+    name: String,
+    allowed_values: String,
+    current_value: String,
+}
+
+impl Display for AuroraScenarioParameter {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}: {} (allowed: {})",
+            self.name, self.current_value, self.allowed_values
+        )
+    }
+}
+
+impl From<aws_sdk_rds::types::Parameter> for AuroraScenarioParameter {
+    fn from(value: aws_sdk_rds::types::Parameter) -> Self {
+        AuroraScenarioParameter {
+            name: value.parameter_name.unwrap_or_default(),
+            allowed_values: value.allowed_values.unwrap_or_default(),
+            current_value: value.parameter_value.unwrap_or_default(),
+        }
+    }
+}
+
+pub struct AuroraScenario {
+    rds: crate::rds::Rds,
+    engine_family: Option<String>,
+    engine_version: Option<String>,
+    instance_class: Option<String>,
+    db_cluster_parameter_group: Option<DbClusterParameterGroup>,
+    db_cluster_identifier: Option<String>,
+    db_instance_identifier: Option<String>,
+    username: Option<String>,
+    password: Option<SecretString>,
+}
+
+impl AuroraScenario {
+    pub fn new(client: crate::rds::Rds) -> Self {
+        AuroraScenario {
+            rds: client,
+            engine_family: None,
+            engine_version: None,
+            instance_class: None,
+            db_cluster_parameter_group: None,
+            db_cluster_identifier: None,
+            db_instance_identifier: None,
+            username: None,
+            password: None,
+        }
+    }
+
+    // Get available engine families for Aurora MySql. rds.DescribeDbEngineVersions(Engine='aurora-mysql') and build a set of the 'DBParameterGroupFamily' field values. I get {aurora-mysql8.0, aurora-mysql5.7}.
+    pub async fn get_engines(&self) -> Result<HashMap<String, Vec<String>>, ScenarioError> {
+        let describe_db_engine_versions = self.rds.describe_db_engine_versions(DB_ENGINE).await;
+        trace!(versions=?describe_db_engine_versions, "full list of versions");
+
+        if let Err(err) = describe_db_engine_versions {
+            return Err(ScenarioError::new(
+                "Failed to retrieve DB Engine Versions",
+                &err,
+            ));
+        };
+
+        let version_count = describe_db_engine_versions
+            .as_ref()
+            .map(|o| o.db_engine_versions().len())
+            .unwrap_or_default();
+        info!(version_count, "got list of versions");
+
+        // Create a map of engine families to their available versions.
+        let mut versions = HashMap::<String, Vec<String>>::new();
+        describe_db_engine_versions
+            .unwrap()
+            .db_engine_versions()
+            .iter()
+            .filter_map(
+                |v| match (&v.db_parameter_group_family, &v.engine_version) {
+                    (Some(family), Some(version)) => Some((family.clone(), version.clone())),
+                    _ => None,
+                },
+            )
+            .for_each(|(family, version)| versions.entry(family).or_default().push(version));
+
+        Ok(versions)
+    }
+
+    pub async fn get_instance_classes(&self) -> Result<Vec<String>, ScenarioError> {
+        let describe_orderable_db_instance_options_items = self
+            .rds
+            .describe_orderable_db_instance_options(
+                DB_ENGINE,
+                self.engine_version
+                    .as_ref()
+                    .expect("engine version for db instance options")
+                    .as_str(),
+            )
+            .await;
+
+        describe_orderable_db_instance_options_items
+            .map(|options| {
+                options
+                    .iter()
+                    .map(|o| o.db_instance_class().unwrap_or_default().to_string())
+                    .collect::<Vec<String>>()
+            })
+            .map_err(|err| ScenarioError::new("Could not get available instance classes", &err))
+    }
+
+    // Select an engine family and create a custom DB cluster parameter group. rds.CreateDbClusterParameterGroup(DBParameterGroupFamily='aurora-mysql8.0')
+    pub async fn set_engine(&mut self, engine: &str, version: &str) -> Result<(), ScenarioError> {
+        self.engine_family = Some(engine.to_string());
+        self.engine_version = Some(version.to_string());
+        let create_db_cluster_parameter_group = self
+            .rds
+            .create_db_cluster_parameter_group(
+                DB_CLUSTER_PARAMETER_GROUP_NAME,
+                DB_CLUSTER_PARAMETER_GROUP_DESCRIPTION,
+                engine,
+            )
+            .await;
+
+        match create_db_cluster_parameter_group {
+            Ok(CreateDbClusterParameterGroupOutput {
+                db_cluster_parameter_group: None,
+                ..
+            }) => {
+                return Err(ScenarioError::with(
+                    "CreateDBClusterParameterGroup had empty response",
+                ));
+            }
+            Err(error) => {
+                if error.code() == Some("DBParameterGroupAlreadyExists") {
+                    info!("Cluster Parameter Group already exists, nothing to do");
+                } else {
+                    return Err(ScenarioError::new(
+                        "Could not create Cluster Parameter Group",
+                        &error,
+                    ));
+                }
+            }
+            _ => {
+                info!("Created Cluster Parameter Group");
+            }
+        }
+
+        Ok(())
+    }
+
+    pub fn set_instance_class(&mut self, instance_class: Option<String>) {
+        self.instance_class = instance_class;
+    }
+
+    pub fn set_login(&mut self, username: Option<String>, password: Option<SecretString>) {
+        self.username = username;
+        self.password = password;
+    }
+
+    pub async fn connection_string(&self) -> Result<String, ScenarioError> {
+        let cluster = self.get_cluster().await?;
+        let endpoint = cluster.endpoint().unwrap_or_default();
+        let port = cluster.port().unwrap_or_default();
+        let username = cluster.master_username().unwrap_or_default();
+        Ok(format!("mysql -h {endpoint} -P {port} -u {username} -p"))
+    }
+
+    pub async fn get_cluster(&self) -> Result<DbCluster, ScenarioError> {
+        let describe_db_clusters_output = self
+            .rds
+            .describe_db_clusters(
+                self.db_cluster_identifier
+                    .as_ref()
+                    .expect("cluster identifier")
+                    .as_str(),
+            )
+            .await;
+        if let Err(err) = describe_db_clusters_output {
+            return Err(ScenarioError::new("Failed to get cluster", &err));
+        }
+
+        let db_cluster = describe_db_clusters_output
+            .unwrap()
+            .db_clusters
+            .and_then(|output| output.first().cloned());
+
+        db_cluster.ok_or_else(|| ScenarioError::with("Did not find the cluster"))
+    }
+
+    // Get the parameter group. rds.DescribeDbClusterParameterGroups
+    // Get parameters in the group. This is a long list so you will have to paginate. Find the auto_increment_offset and auto_increment_increment parameters (by ParameterName). rds.DescribeDbClusterParameters
+    // Parse the ParameterName, Description, and AllowedValues values and display them.
+    pub async fn cluster_parameters(&self) -> Result<Vec<AuroraScenarioParameter>, ScenarioError> {
+        let parameters_output = self
+            .rds
+            .describe_db_cluster_parameters(DB_CLUSTER_PARAMETER_GROUP_NAME)
+            .await;
+
+        if let Err(err) = parameters_output {
+            return Err(ScenarioError::new(
+                format!("Failed to retrieve parameters for {DB_CLUSTER_PARAMETER_GROUP_NAME}"),
+                &err,
+            ));
+        }
+
+        let parameters = parameters_output
+            .unwrap()
+            .into_iter()
+            .flat_map(|p| p.parameters.unwrap_or_default().into_iter())
+            .filter(|p| FILTER_PARAMETER_NAMES.contains(p.parameter_name().unwrap_or_default()))
+            .map(AuroraScenarioParameter::from)
+            .collect::<Vec<_>>();
+
+        Ok(parameters)
+    }
+
+    // Modify both the auto_increment_offset and auto_increment_increment parameters in one call in the custom parameter group. Set their ParameterValue fields to a new allowable value. rds.ModifyDbClusterParameterGroup.
+    pub async fn update_auto_increment(
+        &self,
+        offset: u8,
+        increment: u8,
+    ) -> Result<(), ScenarioError> {
+        let modify_db_cluster_parameter_group = self
+            .rds
+            .modify_db_cluster_parameter_group(
+                DB_CLUSTER_PARAMETER_GROUP_NAME,
+                vec![
+                    Parameter::builder()
+                        .parameter_name("auto_increment_offset")
+                        .parameter_value(format!("{offset}"))
+                        .apply_method(aws_sdk_rds::types::ApplyMethod::Immediate)
+                        .build(),
+                    Parameter::builder()
+                        .parameter_name("auto_increment_increment")
+                        .parameter_value(format!("{increment}"))
+                        .apply_method(aws_sdk_rds::types::ApplyMethod::Immediate)
+                        .build(),
+                ],
+            )
+            .await;
+
+        if let Err(error) = modify_db_cluster_parameter_group {
+            return Err(ScenarioError::new(
+                "Failed to modify cluster parameter group",
+                &error,
+            ));
+        }
+
+        Ok(())
+    }
+
+    // Get a list of allowed engine versions. rds.DescribeDbEngineVersions(Engine='aurora-mysql', DBParameterGroupFamily=<the family used to create your parameter group in step 2>)
+    // Create an Aurora DB cluster database cluster that contains a MySql database and uses the parameter group you created.
+    // Wait for DB cluster to be ready. Call rds.DescribeDBClusters and check for Status == 'available'.
+    // Get a list of instance classes available for the selected engine and engine version. rds.DescribeOrderableDbInstanceOptions(Engine='mysql', EngineVersion=).
+
+    // Create a database instance in the cluster.
+    // Wait for DB instance to be ready. Call rds.DescribeDbInstances and check for DBInstanceStatus == 'available'.
+    pub async fn start_cluster_and_instance(&mut self) -> Result<(), ScenarioError> {
+        if self.password.is_none() {
+            return Err(ScenarioError::with(
+                "Must set Secret Password before starting a cluster",
+            ));
+        }
+        let create_db_cluster = self
+            .rds
+            .create_db_cluster(
+                DB_CLUSTER_IDENTIFIER,
+                DB_CLUSTER_PARAMETER_GROUP_NAME,
+                DB_ENGINE,
+                self.engine_version.as_deref().expect("engine version"),
+                self.username.as_deref().expect("username"),
+                self.password
+                    .replace(SecretString::new("".to_string()))
+                    .expect("password"),
+            )
+            .await;
+        if let Err(err) = create_db_cluster {
+            return Err(ScenarioError::new(
+                "Failed to create DB Cluster with cluster group",
+                &err,
+            ));
+        }
+
+        self.db_cluster_identifier = create_db_cluster
+            .unwrap()
+            .db_cluster
+            .and_then(|c| c.db_cluster_identifier);
+
+        if self.db_cluster_identifier.is_none() {
+            return Err(ScenarioError::with("Created DB Cluster missing Identifier"));
+        }
+
+        info!(
+            "Started a db cluster: {}",
+            self.db_cluster_identifier
+                .as_deref()
+                .unwrap_or("Missing ARN")
+        );
+
+        let create_db_instance = self
+            .rds
+            .create_db_instance(
+                self.db_cluster_identifier.as_deref().expect("cluster name"),
+                DB_INSTANCE_IDENTIFIER,
+                self.instance_class.as_deref().expect("instance class"),
+                DB_ENGINE,
+            )
+            .await;
+        if let Err(err) = create_db_instance {
+            return Err(ScenarioError::new(
+                "Failed to create Instance in DB Cluster",
+                &err,
+            ));
+        }
+
+        self.db_instance_identifier = create_db_instance
+            .unwrap()
+            .db_instance
+            .and_then(|i| i.db_instance_identifier);
+
+        // Cluster creation can take up to 20 minutes to become available
+        let cluster_max_wait = Duration::from_secs(20 * 60);
+        let waiter = Waiter::builder().max(cluster_max_wait).build();
+        while waiter.sleep().await.is_ok() {
+            let cluster = self
+                .rds
+                .describe_db_clusters(
+                    self.db_cluster_identifier
+                        .as_deref()
+                        .expect("cluster identifier"),
+                )
+                .await;
+
+            if let Err(err) = cluster {
+                warn!(?err, "Failed to describe cluster while waiting for ready");
+                continue;
+            }
+
+            let instance = self
+                .rds
+                .describe_db_instance(
+                    self.db_instance_identifier
+                        .as_deref()
+                        .expect("instance identifier"),
+                )
+                .await;
+            if let Err(err) = instance {
+                return Err(ScenarioError::new(
+                    "Failed to find instance for cluster",
+                    &err,
+                ));
+            }
+
+            let instances_available = instance
+                .unwrap()
+                .db_instances()
+                .iter()
+                .all(|instance| instance.db_instance_status() == Some("Available"));
+
+            let endpoints = self
+                .rds
+                .describe_db_cluster_endpoints(
+                    self.db_cluster_identifier
+                        .as_deref()
+                        .expect("cluster identifier"),
+                )
+                .await;
+
+            if let Err(err) = endpoints {
+                return Err(ScenarioError::new(
+                    "Failed to find endpoint for cluster",
+                    &err,
+                ));
+            }
+
+            let endpoints_available = endpoints
+                .unwrap()
+                .db_cluster_endpoints()
+                .iter()
+                .all(|endpoint| endpoint.status() == Some("available"));
+
+            if instances_available && endpoints_available {
+                return Ok(());
+            }
+        }
+
+        Err(ScenarioError::with("timed out waiting for cluster"))
+    }
+
+    // Display the connection string that can be used to connect a 'mysql' shell to the cluster. In Python:
+
+    // Create a snapshot of the DB cluster. rds.CreateDbClusterSnapshot.
+    // Wait for the snapshot to create. rds.DescribeDbClusterSnapshots until Status == 'available'.
+    pub async fn snapshot(&self) -> Result<DbClusterSnapshot, ScenarioError> {
+        warn!("TODO! snapshot");
+        Ok(DbClusterSnapshot::builder().build())
+    }
+
+    pub async fn clean_up(self) -> Result<(), Vec<ScenarioError>> {
+        let mut clean_up_errors: Vec<ScenarioError> = vec![];
+
+        // Delete the instance. rds.DeleteDbInstance.
+        let delete_db_instance = self
+            .rds
+            .delete_db_instance(
+                self.db_instance_identifier
+                    .as_deref()
+                    .expect("instance identifier"),
+            )
+            .await;
+        if let Err(err) = delete_db_instance {
+            let identifier = self
+                .db_instance_identifier
+                .as_deref()
+                .unwrap_or("Missing Instance Identifier");
+            let message = format!("failed to delete db instance {identifier}");
+            clean_up_errors.push(ScenarioError::new(message, &err));
+        } else {
+            // Wait for the instance to delete
+            let waiter = Waiter::default();
+            while waiter.sleep().await.is_ok() {
+                let describe_db_instances = self.rds.describe_db_instances().await;
+                if let Err(err) = describe_db_instances {
+                    clean_up_errors.push(ScenarioError::new(
+                        "Failed to check instance state during deletion",
+                        &err,
+                    ));
+                    break;
+                }
+                let db_instances = describe_db_instances
+                    .unwrap()
+                    .db_instances()
+                    .iter()
+                    .filter(|instance| instance.db_cluster_identifier == self.db_cluster_identifier)
+                    .cloned()
+                    .collect::<Vec<DbInstance>>();
+
+                if db_instances.is_empty() {
+                    trace!("Delete Instance waited and no instances were found");
+                    break;
+                }
+                match db_instances.first().unwrap().db_instance_status() {
+                    Some("Deleting") => continue,
+                    Some(status) => {
+                        info!("Attempting to delete but instances is in {status}");
+                        continue;
+                    }
+                    None => {
+                        warn!("No status for DB instance");
+                        break;
+                    }
+                }
+            }
+        }
+
+        // Delete the DB cluster. rds.DeleteDbCluster.
+        let delete_db_cluster = self
+            .rds
+            .delete_db_cluster(
+                self.db_cluster_identifier
+                    .as_deref()
+                    .expect("cluster identifier"),
+            )
+            .await;
+
+        if let Err(err) = delete_db_cluster {
+            let identifier = self
+                .db_cluster_identifier
+                .as_deref()
+                .unwrap_or("Missing DB Cluster Identifier");
+            let message = format!("failed to delete db cluster {identifier}");
+            clean_up_errors.push(ScenarioError::new(message, &err));
+        } else {
+            // Wait for the instance and cluster to fully delete. rds.DescribeDbInstances and rds.DescribeDbClusters until both are not found.
+            let waiter = Waiter::default();
+            while waiter.sleep().await.is_ok() {
+                let describe_db_clusters = self
+                    .rds
+                    .describe_db_clusters(
+                        self.db_cluster_identifier
+                            .as_deref()
+                            .expect("cluster identifier"),
+                    )
+                    .await;
+                if let Err(err) = describe_db_clusters {
+                    clean_up_errors.push(ScenarioError::new(
+                        "Failed to check cluster state during deletion",
+                        &err,
+                    ));
+                    break;
+                }
+                let describe_db_clusters = describe_db_clusters.unwrap();
+                let db_clusters = describe_db_clusters.db_clusters();
+                if db_clusters.is_empty() {
+                    trace!("Delete cluster waited and no clusters were found");
+                    break;
+                }
+                match db_clusters.first().unwrap().status() {
+                    Some("Deleting") => continue,
+                    Some(status) => {
+                        info!("Attempting to delete but clusters is in {status}");
+                        continue;
+                    }
+                    None => {
+                        warn!("No status for DB cluster");
+                        break;
+                    }
+                }
+            }
+        }
+
+        // Delete the DB cluster parameter group. rds.DeleteDbClusterParameterGroup.
+        let delete_db_cluster_parameter_group = self
+            .rds
+            .delete_db_cluster_parameter_group(
+                self.db_cluster_parameter_group
+                    .map(|g| {
+                        g.db_cluster_parameter_group_name
+                            .unwrap_or_else(|| DB_CLUSTER_PARAMETER_GROUP_NAME.to_string())
+                    })
+                    .as_deref()
+                    .expect("cluster parameter group name"),
+            )
+            .await;
+        if let Err(error) = delete_db_cluster_parameter_group {
+            clean_up_errors.push(ScenarioError::new(
+                "Failed to delete the db cluster parameter group",
+                &error,
+            ))
+        }
+
+        if clean_up_errors.is_empty() {
+            Ok(())
+        } else {
+            Err(clean_up_errors)
+        }
+    }
+}
+
+#[cfg(test)]
+pub mod tests;
