+++ rust_dev_preview/examples/rds/src/aurora_scenario/tests.rs
@@ -1,0 +1,941 @@
+use crate::rds::MockRdsImpl;
+
+use super::*;
+
+use std::io::{Error, ErrorKind};
+
+use assert_matches::assert_matches;
+use aws_sdk_rds::{
+    error::SdkError,
+    operation::{
+        create_db_cluster::{CreateDBClusterError, CreateDbClusterOutput},
+        create_db_cluster_parameter_group::CreateDBClusterParameterGroupError,
+        create_db_instance::{CreateDBInstanceError, CreateDbInstanceOutput},
+        delete_db_cluster::DeleteDbClusterOutput,
+        delete_db_cluster_parameter_group::DeleteDbClusterParameterGroupOutput,
+        delete_db_instance::DeleteDbInstanceOutput,
+        describe_db_cluster_endpoints::DescribeDbClusterEndpointsOutput,
+        describe_db_cluster_parameters::{
+            DescribeDBClusterParametersError, DescribeDbClusterParametersOutput,
+        },
+        describe_db_clusters::{DescribeDBClustersError, DescribeDbClustersOutput},
+        describe_db_engine_versions::{
+            DescribeDBEngineVersionsError, DescribeDbEngineVersionsOutput,
+        },
+        describe_db_instances::{DescribeDBInstancesError, DescribeDbInstancesOutput},
+        describe_orderable_db_instance_options::DescribeOrderableDBInstanceOptionsError,
+        modify_db_cluster_parameter_group::{
+            ModifyDBClusterParameterGroupError, ModifyDbClusterParameterGroupOutput,
+        },
+    },
+    types::{
+        error::DbParameterGroupAlreadyExistsFault, DbClusterEndpoint, DbEngineVersion,
+        OrderableDbInstanceOption,
+    },
+};
+use aws_smithy_http::body::SdkBody;
+use aws_smithy_runtime_api::client::orchestrator::HttpResponse;
+use mockall::predicate::*;
+use secrecy::ExposeSecret;
+
+#[tokio::test]
+async fn test_scenario_set_engine() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster_parameter_group()
+        .with(
+            eq("RustSDKCodeExamplesDBParameterGroup"),
+            eq("Parameter Group created by Rust SDK Code Example"),
+            eq("aurora-mysql"),
+        )
+        .return_once(|_, _, _| {
+            Ok(CreateDbClusterParameterGroupOutput::builder()
+                .db_cluster_parameter_group(DbClusterParameterGroup::builder().build())
+                .build())
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+
+    let set_engine = scenario.set_engine("aurora-mysql", "aurora-mysql8.0").await;
+
+    assert_eq!(set_engine, Ok(()));
+    assert_eq!(Some("aurora-mysql"), scenario.engine_family.as_deref());
+    assert_eq!(Some("aurora-mysql8.0"), scenario.engine_version.as_deref());
+}
+
+#[tokio::test]
+async fn test_scenario_set_engine_not_create() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster_parameter_group()
+        .with(
+            eq("RustSDKCodeExamplesDBParameterGroup"),
+            eq("Parameter Group created by Rust SDK Code Example"),
+            eq("aurora-mysql"),
+        )
+        .return_once(|_, _, _| Ok(CreateDbClusterParameterGroupOutput::builder().build()));
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+
+    let set_engine = scenario.set_engine("aurora-mysql", "aurora-mysql8.0").await;
+
+    assert!(set_engine.is_err());
+}
+
+#[tokio::test]
+async fn test_scenario_set_engine_param_group_exists() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster_parameter_group()
+        .withf(|_, _, _| true)
+        .return_once(|_, _, _| {
+            Err(SdkError::service_error(
+                CreateDBClusterParameterGroupError::DbParameterGroupAlreadyExistsFault(
+                    DbParameterGroupAlreadyExistsFault::builder().build(),
+                ),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+
+    let set_engine = scenario.set_engine("aurora-mysql", "aurora-mysql8.0").await;
+
+    assert!(set_engine.is_err());
+}
+
+#[tokio::test]
+async fn test_scenario_get_engines() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_describe_db_engine_versions()
+        .with(eq("aurora-mysql"))
+        .return_once(|_| {
+            Ok(DescribeDbEngineVersionsOutput::builder()
+                .db_engine_versions(
+                    DbEngineVersion::builder()
+                        .db_parameter_group_family("f1")
+                        .engine_version("f1a")
+                        .build(),
+                )
+                .db_engine_versions(
+                    DbEngineVersion::builder()
+                        .db_parameter_group_family("f1")
+                        .engine_version("f1b")
+                        .build(),
+                )
+                .db_engine_versions(
+                    DbEngineVersion::builder()
+                        .db_parameter_group_family("f2")
+                        .engine_version("f2a")
+                        .build(),
+                )
+                .db_engine_versions(DbEngineVersion::builder().build())
+                .build())
+        });
+
+    let scenario = AuroraScenario::new(mock_rds);
+
+    let versions_map = scenario.get_engines().await;
+
+    assert_eq!(
+        versions_map,
+        Ok(HashMap::from([
+            ("f1".to_string(), vec!["f1a".to_string(), "f1b".to_string()]),
+            ("f2".to_string(), vec!["f2a".to_string()])
+        ]))
+    );
+}
+
+#[tokio::test]
+async fn test_scenario_get_engines_failed() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_describe_db_engine_versions()
+        .with(eq("aurora-mysql"))
+        .return_once(|_| {
+            Err(SdkError::service_error(
+                DescribeDBEngineVersionsError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "describe_db_engine_versions error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let scenario = AuroraScenario::new(mock_rds);
+
+    let versions_map = scenario.get_engines().await;
+    assert_matches!(
+        versions_map,
+        Err(ScenarioError { message, context: _ }) if message == "Failed to retrieve DB Engine Versions"
+    );
+}
+
+#[tokio::test]
+async fn test_scenario_get_instance_classes() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster_parameter_group()
+        .return_once(|_, _, _| {
+            Ok(CreateDbClusterParameterGroupOutput::builder()
+                .db_cluster_parameter_group(DbClusterParameterGroup::builder().build())
+                .build())
+        });
+
+    mock_rds
+        .expect_describe_orderable_db_instance_options()
+        .with(eq("aurora-mysql"), eq("aurora-mysql8.0"))
+        .return_once(|_, _| {
+            Ok(vec![
+                OrderableDbInstanceOption::builder()
+                    .db_instance_class("t1")
+                    .build(),
+                OrderableDbInstanceOption::builder()
+                    .db_instance_class("t2")
+                    .build(),
+                OrderableDbInstanceOption::builder()
+                    .db_instance_class("t3")
+                    .build(),
+            ])
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario
+        .set_engine("aurora-mysql", "aurora-mysql8.0")
+        .await
+        .expect("set engine");
+
+    let instance_classes = scenario.get_instance_classes().await;
+
+    assert_eq!(
+        instance_classes,
+        Ok(vec!["t1".to_string(), "t2".to_string(), "t3".to_string()])
+    );
+}
+
+#[tokio::test]
+async fn test_scenario_get_instance_classes_error() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_describe_orderable_db_instance_options()
+        .with(eq("aurora-mysql"), eq("aurora-mysql8.0"))
+        .return_once(|_, _| {
+            Err(SdkError::service_error(
+                DescribeOrderableDBInstanceOptionsError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "describe_orderable_db_instance_options_error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.engine_family = Some("aurora-mysql".to_string());
+    scenario.engine_version = Some("aurora-mysql8.0".to_string());
+
+    let instance_classes = scenario.get_instance_classes().await;
+
+    assert_matches!(
+        instance_classes,
+        Err(ScenarioError {message, context: _}) if message == "Could not get available instance classes"
+    );
+}
+
+#[tokio::test]
+async fn test_scenario_get_cluster() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .return_once(|_| {
+            Ok(DescribeDbClustersOutput::builder()
+                .db_clusters(DbCluster::builder().build())
+                .build())
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some("RustSDKCodeExamplesDBCluster".to_string());
+    let cluster = scenario.get_cluster().await;
+
+    assert!(cluster.is_ok());
+}
+
+#[tokio::test]
+async fn test_scenario_get_cluster_missing_cluster() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster_parameter_group()
+        .return_once(|_, _, _| {
+            Ok(CreateDbClusterParameterGroupOutput::builder()
+                .db_cluster_parameter_group(DbClusterParameterGroup::builder().build())
+                .build())
+        });
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .return_once(|_| Ok(DescribeDbClustersOutput::builder().build()));
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some("RustSDKCodeExamplesDBCluster".to_string());
+    let cluster = scenario.get_cluster().await;
+
+    assert_matches!(cluster, Err(ScenarioError { message, context: _ }) if message == "Did not find the cluster");
+}
+
+#[tokio::test]
+async fn test_scenario_get_cluster_error() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster_parameter_group()
+        .return_once(|_, _, _| {
+            Ok(CreateDbClusterParameterGroupOutput::builder()
+                .db_cluster_parameter_group(DbClusterParameterGroup::builder().build())
+                .build())
+        });
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .return_once(|_| {
+            Err(SdkError::service_error(
+                DescribeDBClustersError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "describe_db_clusters_error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some("RustSDKCodeExamplesDBCluster".to_string());
+    let cluster = scenario.get_cluster().await;
+
+    assert_matches!(cluster, Err(ScenarioError { message, context: _ }) if message == "Failed to get cluster");
+}
+
+#[tokio::test]
+async fn test_scenario_connection_string() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .return_once(|_| {
+            Ok(DescribeDbClustersOutput::builder()
+                .db_clusters(
+                    DbCluster::builder()
+                        .endpoint("test_endpoint")
+                        .port(3306)
+                        .master_username("test_username")
+                        .build(),
+                )
+                .build())
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some("RustSDKCodeExamplesDBCluster".to_string());
+    let connection_string = scenario.connection_string().await;
+
+    assert_eq!(
+        connection_string,
+        Ok("mysql -h test_endpoint -P 3306 -u test_username -p".to_string())
+    );
+}
+
+#[tokio::test]
+async fn test_scenario_cluster_parameters() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_describe_db_cluster_parameters()
+        .with(eq("RustSDKCodeExamplesDBParameterGroup"))
+        .return_once(|_| {
+            Ok(vec![DescribeDbClusterParametersOutput::builder()
+                .parameters(Parameter::builder().parameter_name("a").build())
+                .parameters(Parameter::builder().parameter_name("b").build())
+                .parameters(
+                    Parameter::builder()
+                        .parameter_name("auto_increment_offset")
+                        .build(),
+                )
+                .parameters(Parameter::builder().parameter_name("c").build())
+                .parameters(
+                    Parameter::builder()
+                        .parameter_name("auto_increment_increment")
+                        .build(),
+                )
+                .parameters(Parameter::builder().parameter_name("d").build())
+                .build()])
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some("RustSDKCodeExamplesDBCluster".to_string());
+
+    let params = scenario.cluster_parameters().await.expect("cluster params");
+    let names: Vec<String> = params.into_iter().map(|p| p.name).collect();
+    assert_eq!(
+        names,
+        vec!["auto_increment_offset", "auto_increment_increment"]
+    );
+}
+
+#[tokio::test]
+async fn test_scenario_cluster_parameters_error() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_describe_db_cluster_parameters()
+        .with(eq("RustSDKCodeExamplesDBParameterGroup"))
+        .return_once(|_| {
+            Err(SdkError::service_error(
+                DescribeDBClusterParametersError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "describe_db_cluster_parameters_error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some("RustSDKCodeExamplesDBCluster".to_string());
+    let params = scenario.cluster_parameters().await;
+    assert_matches!(params, Err(ScenarioError { message, context: _ }) if message == "Failed to retrieve parameters for RustSDKCodeExamplesDBParameterGroup");
+}
+
+#[tokio::test]
+async fn test_scenario_update_auto_increment() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_modify_db_cluster_parameter_group()
+        .withf(|name, params| {
+            assert_eq!(name, "RustSDKCodeExamplesDBParameterGroup");
+            assert_eq!(
+                params,
+                &vec![
+                    Parameter::builder()
+                        .parameter_name("auto_increment_offset")
+                        .parameter_value("10".to_string())
+                        .apply_method(aws_sdk_rds::types::ApplyMethod::Immediate)
+                        .build(),
+                    Parameter::builder()
+                        .parameter_name("auto_increment_increment")
+                        .parameter_value("20".to_string())
+                        .apply_method(aws_sdk_rds::types::ApplyMethod::Immediate)
+                        .build(),
+                ]
+            );
+            true
+        })
+        .return_once(|_, _| Ok(ModifyDbClusterParameterGroupOutput::builder().build()));
+
+    let scenario = AuroraScenario::new(mock_rds);
+
+    scenario
+        .update_auto_increment(10, 20)
+        .await
+        .expect("update auto increment");
+}
+
+#[tokio::test]
+async fn test_scenario_update_auto_increment_error() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_modify_db_cluster_parameter_group()
+        .return_once(|_, _| {
+            Err(SdkError::service_error(
+                ModifyDBClusterParameterGroupError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "modify_db_cluster_parameter_group_error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let scenario = AuroraScenario::new(mock_rds);
+
+    let update = scenario.update_auto_increment(10, 20).await;
+    assert_matches!(update, Err(ScenarioError { message, context: _}) if message == "Failed to modify cluster parameter group");
+}
+
+#[tokio::test]
+async fn test_start_cluster_and_instance() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster()
+        .withf(|id, params, engine, version, username, password| {
+            assert_eq!(id, "RustSDKCodeExamplesDBCluster");
+            assert_eq!(params, "RustSDKCodeExamplesDBParameterGroup");
+            assert_eq!(engine, "aurora-mysql");
+            assert_eq!(version, "aurora-mysql8.0");
+            assert_eq!(username, "test username");
+            assert_eq!(password.expose_secret(), "test password");
+            true
+        })
+        .return_once(|id, _, _, _, _, _| {
+            Ok(CreateDbClusterOutput::builder()
+                .db_cluster(DbCluster::builder().db_cluster_identifier(id).build())
+                .build())
+        });
+
+    mock_rds
+        .expect_create_db_instance()
+        .withf(|cluster, name, class, engine| {
+            assert_eq!(cluster, "RustSDKCodeExamplesDBCluster");
+            assert_eq!(name, "RustSDKCodeExamplesDBInstance");
+            assert_eq!(class, "m5.large");
+            assert_eq!(engine, "aurora-mysql");
+            true
+        })
+        .return_once(|cluster, name, class, _| {
+            Ok(CreateDbInstanceOutput::builder()
+                .db_instance(
+                    DbInstance::builder()
+                        .db_cluster_identifier(cluster)
+                        .db_instance_identifier(name)
+                        .db_instance_class(class)
+                        .build(),
+                )
+                .build())
+        });
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .return_once(|id| {
+            Ok(DescribeDbClustersOutput::builder()
+                .db_clusters(DbCluster::builder().db_cluster_identifier(id).build())
+                .build())
+        });
+
+    mock_rds
+        .expect_describe_db_instance()
+        .with(eq("RustSDKCodeExamplesDBInstance"))
+        .return_once(|name| {
+            Ok(DescribeDbInstancesOutput::builder()
+                .db_instances(
+                    DbInstance::builder()
+                        .db_instance_identifier(name)
+                        .db_instance_status("Available")
+                        .build(),
+                )
+                .build())
+        });
+
+    mock_rds
+        .expect_describe_db_cluster_endpoints()
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .return_once(|_| {
+            Ok(DescribeDbClusterEndpointsOutput::builder()
+                .db_cluster_endpoints(DbClusterEndpoint::builder().status("available").build())
+                .build())
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.engine_version = Some("aurora-mysql8.0".to_string());
+    scenario.instance_class = Some("m5.large".to_string());
+    scenario.username = Some("test username".to_string());
+    scenario.password = Some(SecretString::new("test password".to_string()));
+
+    tokio::time::pause();
+    let assertions = tokio::spawn(async move {
+        let create = scenario.start_cluster_and_instance().await;
+        assert!(create.is_ok());
+        assert_eq!(
+            scenario
+                .password
+                .replace(SecretString::new("BAD SECRET".to_string()))
+                .unwrap()
+                .expose_secret(),
+            &"".to_string()
+        );
+        assert_eq!(
+            scenario.db_cluster_identifier,
+            Some("RustSDKCodeExamplesDBCluster".to_string())
+        );
+    });
+    tokio::time::advance(Duration::from_secs(1)).await;
+    tokio::time::resume();
+    let _ = assertions.await;
+}
+
+#[tokio::test]
+async fn test_start_cluster_and_instance_cluster_create_error() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster()
+        .return_once(|_, _, _, _, _, _| {
+            Err(SdkError::service_error(
+                CreateDBClusterError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "create db cluster error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.engine_version = Some("aurora-mysql8.0".to_string());
+    scenario.instance_class = Some("m5.large".to_string());
+    scenario.username = Some("test username".to_string());
+    scenario.password = Some(SecretString::new("test password".to_string()));
+
+    let create = scenario.start_cluster_and_instance().await;
+    assert_matches!(create, Err(ScenarioError { message, context: _}) if message == "Failed to create DB Cluster with cluster group")
+}
+
+#[tokio::test]
+async fn test_start_cluster_and_instance_cluster_create_missing_id() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster()
+        .return_once(|_, _, _, _, _, _| {
+            Ok(CreateDbClusterOutput::builder()
+                .db_cluster(DbCluster::builder().build())
+                .build())
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.engine_version = Some("aurora-mysql8.0".to_string());
+    scenario.instance_class = Some("m5.large".to_string());
+    scenario.username = Some("test username".to_string());
+    scenario.password = Some(SecretString::new("test password".to_string()));
+
+    let create = scenario.start_cluster_and_instance().await;
+    assert_matches!(create, Err(ScenarioError { message, context:_ }) if message == "Created DB Cluster missing Identifier");
+}
+
+#[tokio::test]
+async fn test_start_cluster_and_instance_instance_create_error() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster()
+        .withf(|id, params, engine, version, username, password| {
+            assert_eq!(id, "RustSDKCodeExamplesDBCluster");
+            assert_eq!(params, "RustSDKCodeExamplesDBParameterGroup");
+            assert_eq!(engine, "aurora-mysql");
+            assert_eq!(version, "aurora-mysql8.0");
+            assert_eq!(username, "test username");
+            assert_eq!(password.expose_secret(), "test password");
+            true
+        })
+        .return_once(|id, _, _, _, _, _| {
+            Ok(CreateDbClusterOutput::builder()
+                .db_cluster(DbCluster::builder().db_cluster_identifier(id).build())
+                .build())
+        });
+
+    mock_rds
+        .expect_create_db_instance()
+        .return_once(|_, _, _, _| {
+            Err(SdkError::service_error(
+                CreateDBInstanceError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "create db instance error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.engine_version = Some("aurora-mysql8.0".to_string());
+    scenario.instance_class = Some("m5.large".to_string());
+    scenario.username = Some("test username".to_string());
+    scenario.password = Some(SecretString::new("test password".to_string()));
+
+    let create = scenario.start_cluster_and_instance().await;
+    assert_matches!(create, Err(ScenarioError { message, context: _ }) if message == "Failed to create Instance in DB Cluster")
+}
+
+#[tokio::test]
+async fn test_start_cluster_and_instance_wait_hiccup() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_create_db_cluster()
+        .withf(|id, params, engine, version, username, password| {
+            assert_eq!(id, "RustSDKCodeExamplesDBCluster");
+            assert_eq!(params, "RustSDKCodeExamplesDBParameterGroup");
+            assert_eq!(engine, "aurora-mysql");
+            assert_eq!(version, "aurora-mysql8.0");
+            assert_eq!(username, "test username");
+            assert_eq!(password.expose_secret(), "test password");
+            true
+        })
+        .return_once(|id, _, _, _, _, _| {
+            Ok(CreateDbClusterOutput::builder()
+                .db_cluster(DbCluster::builder().db_cluster_identifier(id).build())
+                .build())
+        });
+
+    mock_rds
+        .expect_create_db_instance()
+        .withf(|cluster, name, class, engine| {
+            assert_eq!(cluster, "RustSDKCodeExamplesDBCluster");
+            assert_eq!(name, "RustSDKCodeExamplesDBInstance");
+            assert_eq!(class, "m5.large");
+            assert_eq!(engine, "aurora-mysql");
+            true
+        })
+        .return_once(|cluster, name, class, _| {
+            Ok(CreateDbInstanceOutput::builder()
+                .db_instance(
+                    DbInstance::builder()
+                        .db_cluster_identifier(cluster)
+                        .db_instance_identifier(name)
+                        .db_instance_class(class)
+                        .build(),
+                )
+                .build())
+        });
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .times(1)
+        .returning(|_| {
+            Err(SdkError::service_error(
+                DescribeDBClustersError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "describe cluster error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        })
+        .with(eq("RustSDKCodeExamplesDBCluster"))
+        .times(1)
+        .returning(|id| {
+            Ok(DescribeDbClustersOutput::builder()
+                .db_clusters(DbCluster::builder().db_cluster_identifier(id).build())
+                .build())
+        });
+
+    mock_rds.expect_describe_db_instance().return_once(|name| {
+        Ok(DescribeDbInstancesOutput::builder()
+            .db_instances(
+                DbInstance::builder()
+                    .db_instance_identifier(name)
+                    .db_instance_status("Available")
+                    .build(),
+            )
+            .build())
+    });
+
+    mock_rds
+        .expect_describe_db_cluster_endpoints()
+        .return_once(|_| {
+            Ok(DescribeDbClusterEndpointsOutput::builder()
+                .db_cluster_endpoints(DbClusterEndpoint::builder().status("available").build())
+                .build())
+        });
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.engine_version = Some("aurora-mysql8.0".to_string());
+    scenario.instance_class = Some("m5.large".to_string());
+    scenario.username = Some("test username".to_string());
+    scenario.password = Some(SecretString::new("test password".to_string()));
+
+    tokio::time::pause();
+    let assertions = tokio::spawn(async move {
+        let create = scenario.start_cluster_and_instance().await;
+        assert!(create.is_ok());
+    });
+
+    tokio::time::advance(Duration::from_secs(1)).await;
+    tokio::time::advance(Duration::from_secs(1)).await;
+    tokio::time::resume();
+    let _ = assertions.await;
+}
+
+#[tokio::test]
+async fn test_scenario_clean_up() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_delete_db_instance()
+        .with(eq("MockInstance"))
+        .return_once(|_| Ok(DeleteDbInstanceOutput::builder().build()));
+
+    mock_rds
+        .expect_describe_db_instances()
+        .with()
+        .times(1)
+        .returning(|| {
+            Ok(DescribeDbInstancesOutput::builder()
+                .db_instances(
+                    DbInstance::builder()
+                        .db_cluster_identifier("MockCluster")
+                        .db_instance_status("Deleting")
+                        .build(),
+                )
+                .build())
+        })
+        .with()
+        .times(1)
+        .returning(|| Ok(DescribeDbInstancesOutput::builder().build()));
+
+    mock_rds
+        .expect_delete_db_cluster()
+        .with(eq("MockCluster"))
+        .return_once(|_| Ok(DeleteDbClusterOutput::builder().build()));
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("MockCluster"))
+        .times(1)
+        .returning(|id| {
+            Ok(DescribeDbClustersOutput::builder()
+                .db_clusters(
+                    DbCluster::builder()
+                        .db_cluster_identifier(id)
+                        .status("Deleting")
+                        .build(),
+                )
+                .build())
+        })
+        .with(eq("MockCluster"))
+        .times(1)
+        .returning(|_| Ok(DescribeDbClustersOutput::builder().build()));
+
+    mock_rds
+        .expect_delete_db_cluster_parameter_group()
+        .with(eq("MockParamGroup"))
+        .return_once(|_| Ok(DeleteDbClusterParameterGroupOutput::builder().build()));
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some(String::from("MockCluster"));
+    scenario.db_instance_identifier = Some(String::from("MockInstance"));
+    scenario.db_cluster_parameter_group = Some(
+        DbClusterParameterGroup::builder()
+            .db_cluster_parameter_group_name("MockParamGroup")
+            .build(),
+    );
+
+    tokio::time::pause();
+    let assertions = tokio::spawn(async move {
+        let clean_up = scenario.clean_up().await;
+        assert!(clean_up.is_ok());
+    });
+
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for first Describe Instances
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for second Describe Instances
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for first Describe Cluster
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for second Describe Cluster
+    tokio::time::resume();
+    let _ = assertions.await;
+}
+
+#[tokio::test]
+async fn test_scenario_clean_up_errors() {
+    let mut mock_rds = MockRdsImpl::default();
+
+    mock_rds
+        .expect_delete_db_instance()
+        .with(eq("MockInstance"))
+        .return_once(|_| Ok(DeleteDbInstanceOutput::builder().build()));
+
+    mock_rds
+        .expect_describe_db_instances()
+        .with()
+        .times(1)
+        .returning(|| {
+            Ok(DescribeDbInstancesOutput::builder()
+                .db_instances(
+                    DbInstance::builder()
+                        .db_cluster_identifier("MockCluster")
+                        .db_instance_status("Deleting")
+                        .build(),
+                )
+                .build())
+        })
+        .with()
+        .times(1)
+        .returning(|| {
+            Err(SdkError::service_error(
+                DescribeDBInstancesError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "describe db instances error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    mock_rds
+        .expect_delete_db_cluster()
+        .with(eq("MockCluster"))
+        .return_once(|_| Ok(DeleteDbClusterOutput::builder().build()));
+
+    mock_rds
+        .expect_describe_db_clusters()
+        .with(eq("MockCluster"))
+        .times(1)
+        .returning(|id| {
+            Ok(DescribeDbClustersOutput::builder()
+                .db_clusters(
+                    DbCluster::builder()
+                        .db_cluster_identifier(id)
+                        .status("Deleting")
+                        .build(),
+                )
+                .build())
+        })
+        .with(eq("MockCluster"))
+        .times(1)
+        .returning(|_| {
+            Err(SdkError::service_error(
+                DescribeDBClustersError::unhandled(Box::new(Error::new(
+                    ErrorKind::Other,
+                    "describe db clusters error",
+                ))),
+                HttpResponse::new(SdkBody::empty()),
+            ))
+        });
+
+    mock_rds
+        .expect_delete_db_cluster_parameter_group()
+        .with(eq("MockParamGroup"))
+        .return_once(|_| Ok(DeleteDbClusterParameterGroupOutput::builder().build()));
+
+    let mut scenario = AuroraScenario::new(mock_rds);
+    scenario.db_cluster_identifier = Some(String::from("MockCluster"));
+    scenario.db_instance_identifier = Some(String::from("MockInstance"));
+    scenario.db_cluster_parameter_group = Some(
+        DbClusterParameterGroup::builder()
+            .db_cluster_parameter_group_name("MockParamGroup")
+            .build(),
+    );
+
+    tokio::time::pause();
+    let assertions = tokio::spawn(async move {
+        let clean_up = scenario.clean_up().await;
+        assert!(clean_up.is_err());
+        let errs = clean_up.unwrap_err();
+        assert_eq!(errs.len(), 2);
+        assert_matches!(errs.get(0), Some(ScenarioError {message, context: _}) if message == "Failed to check instance state during deletion");
+        assert_matches!(errs.get(1), Some(ScenarioError {message, context: _}) if message == "Failed to check cluster state during deletion");
+    });
+
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for first Describe Instances
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for second Describe Instances
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for first Describe Cluster
+    tokio::time::advance(Duration::from_secs(1)).await; // Wait for second Describe Cluster
+    tokio::time::resume();
+    let _ = assertions.await;
+}
